+++ docs/changelog.md
@@ -10,7 +10,7 @@ ITables ChangeLog
 - ITables is now tested with Python 3.12
 
 **Changed**
-- ITables uses the latest version `2.0.2` of `datatables-net` ([#121](https://github.com/mwouts/itables/issues/121))
+- ITables uses the latest version `2.0.2` of `DataTables` ([#121](https://github.com/mwouts/itables/issues/121))
 - Large Python integers are now mapped to JavaScript `BigInt` ([#172](https://github.com/mwouts/itables/issues/172))
 - ITables is build using `hatch` and `pyproject.toml`